Refactor commit components
